% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tree_status.R, R/cleaningPlots.R
\name{classify_tree_status}
\alias{classify_tree_status}
\title{classify trees according to their status (living, dead or regenerated)}
\usage{
classify_tree_status(Trees)

classify_tree_status(Trees)
}
\arguments{
\item{Trees}{a data.table of individual tree measurements within species (PSPs)}
}
\value{
a list of trees categorized as Regeneration, Mortality, or Alive

a list of trees categorized as Regeneration, Mortality, or Alive
}
\description{
classify trees according to their status (living, dead or regenerated)

classify trees according to their status (living, dead or regenerated)
}
