% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R, R/dbh_check.R
\name{process_dbh_issues}
\alias{process_dbh_issues}
\title{processing problems with dbh in PSP data}
\usage{
process_dbh_issues(Trees, plots)

process_dbh_issues(Trees, plots)
}
\arguments{
\item{Trees}{a data.table of individual tree measurements within species (PSPs)}

\item{plots}{a vector of plot IDs}
}
\value{
a list of standardized plot and tree data.tables

a list of standardized plot and tree data.tables
}
\description{
processing problems with dbh in PSP data

processing problems with dbh in PSP data
}
