% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningPlots.R,
%   R/correction_a_treenum_to_multiplePSP.R
\name{treenum_to_multiplePSP}
\alias{treenum_to_multiplePSP}
\title{Finding where multiple PSPs are associated with the same treenum in the same Plot}
\usage{
treenum_to_multiplePSP(Trees)

treenum_to_multiplePSP(Trees)
}
\arguments{
\item{Trees}{a data.table of individual tree measurements within species (PSPs)}
}
\value{
a list containing corrected PSP assignments

a list containing corrected PSP assignments
}
\description{
Finding where multiple PSPs are associated with the same treenum in the same Plot

Finding where multiple PSPs are associated with the same treenum in the same Plot
}
